refactor: add ai folder
