+++ src/ai/endgameTablebase.ts
@@ -1,6 +1,6 @@
 // src/endgameTablebase.ts
-import { Board } from './board';
-import { PieceColor, PieceType } from './piece';
+import { Board } from '../board';
+import { PieceColor, PieceType } from '../piece';
 
 type Move = { fromX: number; fromY: number; toX: number; toY: number };
